export resolver and reorganize
