+++ libs/category/src/resolvers/public_api.ts
@@ -1,0 +1,2 @@
+export { DaffDefaultCategoryPageSize } from './category-page-server-side/default-category-page-size.token';
+export { DaffCategoryPageServerSideResolver } from './category-page-server-side/category-page-server-side.resolver';
